match: Move match card and header component to match folder
